+++ database/migrations/2023_04_24_230256_create_media_table.php
@@ -1,0 +1,38 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('media', function (Blueprint $table) {
+            $table->id();
+            $table->string('filename');
+            $table->string('original_filename');
+            $table->string('mime_type');
+            $table->string('path');
+            $table->string('disk')->default('public');
+            $table->unsignedBigInteger('size');
+            $table->string('collection')->default('default');
+            $table->longText('alt_text')->nullable();
+            $table->longText('caption')->nullable();
+            $table->json('metadata')->nullable();
+            $table->string('created_by')->nullable();
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('media');
+    }
+};